refactor: Command 定義を Actor object に移動
typed Actor の標準に合わせるため
