+++ payment-app/application/src/main/resources/reference.conf
@@ -54,7 +54,7 @@ akka.actor {
   }
   serialization-bindings {
     "jp.co.tis.lerna.payment.application.ecpayment.issuing.actor.ECPaymentIssuingServiceEvent" = kryo
-    "jp.co.tis.lerna.payment.application.ecpayment.issuing.actor.Command" = kryo
+    "jp.co.tis.lerna.payment.application.ecpayment.issuing.actor.PaymentActor$Command" = kryo
     "jp.co.tis.lerna.payment.adapter.ecpayment.issuing.model.SettlementResponse" = kryo
 
     "jp.co.tis.lerna.payment.application.util.shutdown.TryGracefulShutdown" = kryo
